requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
face-recognition==1.3.0
numpy==1.26.2
Pillow==10.1.0
jinja2==3.1.2
